+ systune: sysctl: ping_group_range: allow all system gids to use ping
somewhere along the way, the issue of ping being setuid root (to allow
users to make raw sockets) became seen as a security issue.  so we
started putting cap_net_raw on /bin/ping.  but these capabilities bits
were getting dropped or never enabled for various reasons so SOCK_DGRAM
is now used with IPPROTO_ICMP and a tunable in
/proc/sys/net/ipv4/ping_group_range says what system users can make
these kinds of sockets.  we see no reason to restrict ping so we'll open
it to the entire gid space.  maximum upper limit was determined
empirically.  upstream should have used 0 or -1 to accommodate the
future but, as it is implemented today, this range is max.