+ tf: vars: imgtypes: add the kube node type
we were already operating with this on the deployed one, this is just
the barebones template.  probably we need to encrypt this file and
change it, rather than have a "checked out" dirty copy, but currently it
changes too much for lab tests, etc and would be too noisy to have in
git, so we just have an example "template" of the file's contents.  but
should at least account for all the node types properly...