+ tf: kubestrap: use ansible rather than direct provision new kube nodes
by having the provisioning defined in ansible, we can do a lot more
things on the new node and have a lot better syntax and expansions
available.

an initial motivation for this was, we won't be installing flannel
anymore by "kubectl apply" but rather using helm we will install cilium,
and we didn't want to be doing this with a bunch of serial local-execs.
the set of things to do on a newly-provisioned master (and possibly even
slave, although those should just fall in line mostly) will only grow
with time.

so now we have just formalized the post-instantiation provisioning out
to an ansible playbook.  we will thus have more flexibility and can use
the ansible kubernetes collection to do a bunch of stuff to get the
cluster going.