+ tfapply: exit_handler: add print of error from jsonfile if bomb
if we're exiting with an error status, we already know the output
filenames, so test the first one that's non-empty and print the
diagnostic fields from within.  these seem to be the ones that contain
the actual error messages.  this works for the use cases we've
identified but we'll see if it holds for other use cases as we continue
to use the wrapper.

this is very helpful because we had to do a ls -lt, cat as root, parse
with jq, and page it, then skip to the error location.  this happened
too many times and we think this is better done automatically.