+ kubestrap: install cilium repo/deployment via inline helm values
instead of flannel (which we prototyped with), cilium is a much more
complete, capable and performant system.  our install:

  - takes over for kube-proxy
  - sets cluster name/id according to podnet/plex (for clustermesh)
  - specifies the podnet in ipam config
  - features:
    - gatewayapi
    - egress gateway
    - dedicated load balancer
    - native routing mode
    - bpf masquerading
    - direct layer 2 announcements

hubble disabled for now, commented out block about it being enabled, we
have to disable because it requires dataplane nodes to be up, there is
no tolerance for control-plane taint.  so hubble has to come after
cluster bootstrap has finished and we have added some dataplane nodes.