+ tf: main: plexhocs: lifecycle: make ignore of image, config shorter
no need to span this on multiple lines and put specific keys.  we know
what this is for, is obvious from looking at it: allow the config and
backing images to change (get updated) without forcing a replacement of
deployed nodes.  we can destroy them with -target if we want that.