- tf: locals: replace data.external.env.HOME with pathexpand("~")
there is no need to use the environment variable when we have
pathexpand, this is one less variable we'll have to pass in template
arguments.  the data reference's string as an interpolated value happens
to be the exact same length as using the textual string for the
pathexpand() call also, so it doesn't change any string lengths.  it
just will let us move cloudinit templating to a module and then not have
to pass the home page or set up a duplicate data source within the
module as we already had in root module.