+ tfapply: validate: do a validation pass and display stdout on failure
plan already validates, but we use json output for that, and in the case
of validate, we prefer the tool-generated output, it usually has a code
block, line number and colorization, which we normally don't want for
applies, but we do want if validation fails.  so stash the output in
another temporary file and display it on nonzero exit of the validate
subcommand.