- tf: cloudinit: templatefile: remove some args and do it in the template
we are refactoring because we need to pass less args to the
templatefile, so we can use it in multiple resources.  if there are too
many arguments it will be unwieldy.  so these things that can have more
or less hardcoded expansions, we can move inside the template itself.
we might create a module for this and move these again to be in the
module locals.