+ tf: netconfig: make it a template file of its own, and a template dir
we're moving the network-config into its own template file to reduce the
amount that we have to copy to duplicate the plexhoc resource (one for
kubemaster, one for kubeworkers).

now that we have two templates, and could have more in the future, then
rather than clutter the *.tf files with several *.tftpl files, we'll
also make a dedicated tftpl/ directory inside the terraform/ dir and put
all our templates there.