= tf: fmt: update some but not all tofu fmt changes
some of the changes "tofu fmt" advises make no sense, using indents (or
usually lack of indents) that make it difficult to read.  we're the only
one using this repo anyways.  we will continue to use most but not all
the fmt suggestions and the ones we don't use we're only typically off
by one indent.