+ tf: cloudinit: new module that encapsulates cloudinits for each node
we have to use this for multiple resources and did not want to have to
make the huge templatefile() call with its large parameter list multiple
times when only a few key differences need to be provided, in particular
the each.key.  most things can be calculated within the module, but
we have to move a lot of logic that's only used by the cloudinit
templates (right now) into the module.  after doing so it simplifies
things considerably, as we can make those variables that are global into
module arguments and within the module, make a single large data
structure with all nodes' cloudinits.  then index into it with the node
name at the cloudinit usage sites.  this greatly simplifies using the
cloudinits for root module resources.