- vars: pysetup: hardcode python path var rather than discover
when we need the setup users's interpreter (for all its libraries),
rather than the usual config user's, we just hardcode the path -- which
won't really change, and if it does we can do it in one place -- rather
than discover it via PATH and then have to do it in every play that uses
it.  if we put the libraries in config user python (ie, using
python3/cfgnode roles), then all systems need to have every library.  in
most cases these are for running tasks on the controller (such as for
kubernetes tasks when we use the local kubeconfig) and we just want to
use the same interpreter that ansible controller uses, that's where we
manage the requirements.txt with the kitchen sink and that's where we
run control-plane tasks.