- tf: main: plexhocs: remove running=true since this is default
no need to specify this since that is the default.  right now we're
trying to reduce the size of the config stanza because we'll have
multiple resources with this whole block, which unfortunately cannot be
comnbined but must be copied.

probably we had left this "running=true" in so we could know that it was
there, since when we bake, we need a running=false, and require to
specify a no-refresh to do it.  but we remember this by now and will fix
that image bake issue (which requires us to have a custom version of
terraform collection) separately.