+ kubestrap: wait for node-ready so tf pvn knows when join complete
this way terraform won't proceed with any subsequent steps (dependent
resources) until the node is actually ready and has joined the cluster.
we were already doing that for kubemasters indirectly with cilium wait
(which we left in place), but now it will work on slave nodes as well.

tried using only this readiness check, and removing cilium wait, but
actually the node is not quite finished setting up the other cilium pods
yet, only the base pod being created is already enough to make the node
ready in the kubernetes sense (ie it can start scheduling other stuff...
which, indeed, it does, to start the remaining cilium pods like envoy).
so for the controller, we still leave the "cilium status --wait" in
place.  for this case, actually the node readiness check is redundant,
but also harmless.