+ tf: plexnodes: separate resources for kubemasters, kubeslaves, plexnons
we have different needs for these and unfortunately we cannot have
self-referential dependencies for a generic resource block which require
an expression using each.key or each.value to calculate, because
depends_on does not support expressions.  we also tried using a
terraform_data resource indirection, and this has the same result, a
cyclical dependency.  after making several attempts and reading many
github issues and forum posts, decided that it just wasn't ever going to
work, we need to have separate resources.  unfortunately this requires a
lot of copying of identical stanza, but there's no way to avoid this.
we can try moving some/all of these differences into a module later.

for kubeslaves, we implemented a destroy-time provisioner that drains
and removes the node from the cluster and wraps the provisioners in an
external script so we can (1) do more things, (2) refer to variables,
which is not allowed in destroy-time provisioners in terraform (although
it used to be, but was removed because it apparently causes issues...
see ticket referenced in comments).

not really happy with verbosity of the resulting 3x stanza but it's the
best we can do for now and does have the required behavior that we can
pick nodes and reprovision them, and recreating the master recreates the
whole thing.  this will let us re-instantiate the whole cluster by just
replacing the master node.

we will next add the terraform flux provider, so we can completely
re-instantiate the whole cluster and have it build itself to the same
exact state just by deleting the master node.

TODO: supporting multiple master control-plane nodes, unnecessary for
our lab system where each "plex" runs on a single machine, but let's see
what the future will bring...