/ tf: kubeslaves: trigger slave replacements when kubemaster changes
we had the depends set up properly, so they were getting sequenced
correctly, and the master node had to be up before slaves would proceed
to provision themselves.  but then when we replaced the master, there
was no change on the slaves, which won't work, because the keys would be
different and the slaves would be orphaned members of the previous
cluster that no longer exists.  so we have to be explicit about the
replacement trigger just as we do with the depends trigger.

note that this may or may not work with multi-masters.  for example if
we just added or removed master nodes, would that cause all slaves to be
replaced? which it should not, as long as at least one kube-apiserver
remains then the control plane would all be the same even if we scaled
the number of masters. but right now we only support one master so this
is not something we have to think about.  likely we would use a
terraform_data resource indirection that had an input that did some
calculation or referred to the cluster ID or something like that.  we'll
worry about this later, for now we want (1) correct bootstrap ordering,
(2) reproducibility, (3) immutability and (4) properly move on to the
application layer bootstrap, with flux.