+ tf: main: profile: knode: lxc.log.level make it 1 which is debug level
they aren't really voluminous at all and there is just a single file in
/var/log/incus/<project>_<nodename>/

actually we should do this on all of them, but when we went to do that,
discovered that profiles do not merge keys, so we have to repeat it in
leaves.  this is sad but we'd need an indirection somewhere to
encapsulate it.