/ kubenode: overlay empty tmpfs on drm, graphics, and input for udev
we do not want udev scan/adds in the system container to generate
reconnect events (for example disrupting Xorg and causing a flurry of
module reloads and re-init code being run, like for keyboard and mouse)
so we add an overlay in /sys to an empty tmpfs mount.  udev will not see
the devices there and emit its add events.  hack, but works for now.