+ tf: registryimg: new incus image resource with docker registry oci
this defines a docker hub image of
https://github.com/distribution/distribution/ which is the docker
registry prorgram in golang, split into a separate project.  it can run
in simple app container with configuration from environment variable.

we will add the instance that uses it in upcoming commits.

note that we add the remote here.  currently the commented-out details
are required to be in the local ~/.config/incus/config.yml, but this is
bug according to upstream and fix is pending.  we will uncomment those
and switch to it in a later commit.