+ tf: plexocireg: new incus instance, uses registryimg oci img on plex ip #2
we have previously committed hosts file change to make all plexes have a
plexname-oci as second ip after the plex ip.  on those we run an oci
image that listens on 5000 running an oci registry in so-called
"pull-through cache" mode.  using the official docker hub oci image of
this.  the project itself is the docker registry, now a standalone
project, to be found at https://github.com/distribution/distribution/
and since the OCI support in incus is minimal, we have to do some things
in the shell when instantiating this.  in particular it's using some
kind of forknet thing to give it dhcp, it bind mounts all the files in
/var/lib/incus/containers/*/network/, it looks to me like this is what
always happens, but the images for ubuntu etc that boot in the system
containers actually do this same thing where they umount -l and create
the resolv.conf.  the problem is that the bind mount of /etc/resolv.conf
gets the copy from the host, which lists different things on different
plex hosts and aren't always accessible from a plexnode.  we want to
create it.

and then forcing the static ip... this is really the only way to do it
since there's nothing on the host that does anything like that.  it
would get dhcp-assigned address and route but that's about it.