+ kubenode: localize_oci_registry_upstream: bake site oci cache into images
now that we have an incus_instance running an oci of a local docker
registry instance in pull-through cache mode, running on
plexname-oci:5000 on each plexhost, we will now bake this into images.
it's necessary to get this into the crio configuration because the
initial bootstrap from kubeadm will try to download it.  we cannot use a
persistent store on the nodes, because, while this works, it's useless
for bootstrap scenario since the node is created from scratch, it will
never be able to cache anything.  cache needs to exist outside the
bootrap, this is the purpose of our caching proxy incus_instance