- kubefacts: factor out node type seek into task file, to be sourced
we will centralize a few things, in particular it's good to determine
the type of node at runtime, we don't yet have groups for this, but the
process table is reliable in this regard, at least to introspect the
running system and derive the information.

we anticipate other users of this information upcoming, so we will
factor this out into its own file.