+ tfapply: display partially interpreted, json plan error response
the actual error beyond summary doesn't seem to appear on stdout or
stderr during the plan phase, so we are forced to use the json itself.
we have filtered out fields we don't care about, mainly error range
start and end cols, and end of anything.  we have kept only start line
and col which should be enough if the user will really be tracking it
down, to load in the editor.  other than that we keep jq for some and
display any top-level strings without their name since that will be
obvious from the message.  it results in a decent split between concise
and easy to scan.  the actual rendered stderr is better, some commands
seem to keep it (like apply) and we don't have to parse the json.  but
plan seems to be special in that it won't display the pretty stderr
(maybe if output is not a terminal?)