- tfapply: remove blank lines from output, now have to force-colorize
we want to get rid of the blank lines which happen from the printing of
empty strings by jq pretty-printer.  sed can do this quite handily, but
then output is no longer going to a terminal, so we have to
force-colorize with jq.  this is an interactive-only tool without an
intended scripting use case, so we can assume for the time being that
it's interactive-only.