- kubestrap: wait_for_knode_readiness: allow six minute bootstrap timeout
yes, bootstrap has taken this long before, there seems to be a large
variation in the speed of upstream OCI registries.  we do have the
Zot registry cache now, which helps for the repeat bootstraps, but
there's still the initial one, which potentially hasn't happened in a
while and versions have changed, caches have been purged, etc so
sometimes the longer downloads do have to happen and we can't timeout.
actually 4 minutes is more typical but 6 minutes plus has happened, so
we may actually end up increasing this again at some point...