/ tf: plexocireg: set preserveDigest so docker2s2 mode works
this took a while to figure out, basically Zot is OCI-only but has a
Docker compatibility mode that tries to convert Docker-metadata images
to OCI on the fly but then that changes digests, which screws everything
up and there's a bunch of retries and redos and iterating on kube
bootstrap ends up taking forever.  stumbled on this fix in
project-zot/zot issue 3295.

with this fix, the registry cache works *much* faster, than even
disabling docker2s2.  so many software pipelines are still spitting out
Docker images, apparently.