+ tf: knodes: new module, per-plex kube masters and slaves with dependencies
here we have generalized the kubernetes nodes into a module and now have
both master and slave module instances for each of the plex hosts.  the
slave module instance relies on the master with depends_on, which
confers the instantiation order, and then in the master module instance
[only] we output the "primary" master instance.  this is then relied on by a
terraform_data instance in the slave module instance, which lets us
effectively wrap an expression (in the terraform_data) in a resource so
it can be used within a lifecycle replace_triggered_by (which does not
allow expressions).  this way we can (1) assure that the master comes up
before any slaves; (2) replacing the master will cascade to the slaves,
which of course depend on the master and require a reset and
re-bootstrap if it changes (as immutable servers these would always be
rebuilt rather than try to rejoin a new cluster or something like that)

in theory there could be multiple masters, and it was sort of written in
a way to extend it for that later, but this probably won't work right if
we're multi-master now.  however, we aren't yet, and don't need to worry
about that.  these plexes are always single [physical] machines anyways.