/ tf: knodes: master: properly rely on mac_address attribute
we had the mistaken impression that we can create multiple outputs with
different field members of inputs map, but actually a single
terraform_data only supports one input.  if it's a composite input (ie,
a map with multiple variable inputs), it's still a single variable in
the state, and thus it will always change if any of the composite
members are changed, because the whole thing changed.