- tf: knodes: terraform_data: master: rename resource, from kubedata
since we can only encode one variable, it's not a bunch of data about
the kube, but one specific datum, namely the master node.  and we
started with masterid (it has a mac address), but really we want an
opaque identifier, we already know it's a terraform_data output.  we'll
see if we change our mind later about the name.