tf: knodes: terraform_data: master: rm comment about count, not possible
the reason it's not possible is not unknown, it's because it's used as a
replace_triggered_by and therefore has to always exist.  will continue
to ponder on this if there's some way to make it conditional on actually
being a slave, where the information is relevant (masters already know
about themselves).